Image transpositions and plotting improvments
The txtToMat.m function implicitly applies a transposition to the input data; I now apply an inverse transposition in this function to bring the read images in line with the output of Gwyddion, and have removed a number of ad hoc transpositions in the visualisation code.

This update also adds the option to plot the mass distribution of fibre orientations (essentially a length-weighted histogram) and applies a couple of fixes to the network analysis stages.